refactor : разложил по папкам, подправил зависимости
